Created an intermediate class for sweeped knobs.
This class handles knobs with sweep: the number of degrees of motion it
has, centered at 12 o'clock. (The equivalent of the total mechanical
travel in a potentiometer.)  Both rotary knobs and center-tapped knobs
are of this type; continuously rotating knobs are not.

Also refactored some code, so that the methods are in the proper order,
and corrected the relevant JavaDoc.