Removed debug messages
FUCKING GIT. I accidentally committed Knob.java, which I did NOT want to
do. Unfortunately, it is impossible to revert that commit without also
overwriting all the changes that I made.

So, for now, I am committing the (broken) Knob.java code to the repo,
and removing the debug messages. For now.